refactor: create authentication helper
