+++ tests/integration/local/conftest.py
@@ -10,7 +10,7 @@ from jina import Flow
 
 from now.admin.utils import get_default_request_body
 from now.common.options import construct_app
-from now.constants import Apps, DatasetTypes, Models, S3_CUSTOM_MM_DATA_PATH
+from now.constants import S3_CUSTOM_MM_DATA_PATH, Apps, DatasetTypes, Models
 from now.data_loading.create_dataclass import create_dataclass
 from now.data_loading.data_loading import load_data
 from now.demo_data import DemoDatasetNames
@@ -20,7 +20,7 @@ from now.executor.autocomplete import NOWAutoCompleteExecutor2  # noqa: F401
 from now.executor.gateway import NOWGateway  # noqa: F401
 from now.executor.indexer.elastic import NOWElasticIndexer  # noqa: F401
 from now.now_dataclasses import UserInput
-from now.utils import get_aws_profile, write_flow_file
+from now.utils.authentication.helpers import get_aws_profile, write_flow_file
 
 BASE_URL = 'http://localhost:8081/api/v1'
 SEARCH_URL = f'{BASE_URL}/search-app/search'
